Added two methods to the ExpenseRecord class, add_new_record and remove_record.
add_new_record(expense_ws) takes the 'Expense' worksheet as an argument and
inserts the new record at the top.

remove_record(expense_ws, row_num) takes the 'Expense' worksheet and row number
as arguments.
(As this program is working with a formatted worksheet, the first 4 rows are
reserved for title header and field headers and will need to be considered
when deleting records)

To remove a record from the worksheet:
	1) Instantiate new ExpenseRecord object with no arguments
	2) Call the remove_record method that belongs to this object
	   while passing expense_ws and row_num as arguments
	3) The method will set the instance variables using the values from
	   the row, before deleting the row and returning the updated record object