Lots of clean up
Cleaned out the main.py file. As previously stated, I moved all excel related code into a separate file.
This excel related code will be a feature of the overall project.

Will now return to the actual purpose of this branch, the GUI.

The layout will be (subject to change, but the base idea will remain the same):

Main view with 3 tabs:
	Income
	Expense
	Budget

Each tab will display a separate frame in the main view, only the active frame will be displayed:

Income and Expense components:
	1)  Central frame with records, not unlike the cell style in a spreadsheet
		-will display Date, Expense/Income description, expense/income amount
	2)  Buttons to add, delete, update records.
		-adding and deleting will open a new window with input boxes to enter/edit values
	3)  A frame to display total income/expense
	4)  Also a date range selector to get records and display totals based on a given range of dates
		-should default to the last 30 days. Or maybe to the current month with an option to
		display the last 15/30/60 days, not unlike what's seen in online banking

Balance components:
	1)  Total income
        2)  Total expense
	3)  Total balance
	4)  Date selector like above
	5)  Bonus: After MVP, add projected balance calculations based off 15/30/60 day averages
		    or some other calculation I will look up