Changed out radio selector for actual buttons to change between Income, Expense, and Balance frames
The buttons work and look decent enough for now. The buttons behave as they should, visually as well as
functionally.

Trimmed down the lines of code by using a decorator show_active_frame() to add to the behavior of the button
callback functions. Was fun, never used decorators before. Will be using and abusing that feature in this language
at any reasonable opportunity.

GUI is coming along. Need to design the record frames that will populate the Income and Expense Frames

Considering ditching the balance frame. Perhaps display the basics of the balance in the lower part of the main view
with a separate button to open a new window for more detail.
Or maybe a menu to select between views on the main window.
Time will tell.